reduced firmware size to fit pro micro
