+++ keyboards/lotus58/keymaps/via/rules.mk
@@ -2,3 +2,7 @@ OLED_DRIVER_ENABLE = yes
 WPM_ENABLE = yes
 LTO_ENABLE = yes
 VIA_ENABLE = yes
+FAUXCLICKY_ENABLE = no
+HD_44780_ENABLE = no
+
+
