+++ keyboards/lotus58/rev1/config.h
@@ -48,3 +48,8 @@
 
 #define NO_ACTION_MACRO
 #define NO_ACTION_FUNCTION
+
+#define NO_ACTION_ONESHOT
+#define NO_ACTION_TAPPING
+#define IGNORE_MOD_TAP_INTERRUPT
+#define TAPPING_FORCE_HOLD
