+++ keyboards/lotus58/rules.mk
@@ -11,7 +11,6 @@ NKRO_ENABLE = no
 AUDIO_ENABLE = no
 UNICODE_ENABLE = no
 BLUETOOTH_ENABLE = no
-RGBLIGHT_ENABLE = yes
 
 ENCODER_ENABLE = yes
 EXTRAFLAGS += -flto
